Improve CALL-WITH-OVERLAPPING for GRID structure
Expand search region in geometric space instead of index space. The
benefit is that we can detect when the query region is entirely
outside of the volume covered by the grid.

Before this change, the clamping in index space lead to at least one
one cell position for each dimension being searched.