Fix overlapping and containment tests for objects in quadtree
Before this change, the functions REGION-{CONTAINS,OVERLAPS}-AREA-P
received a location and a size which they treated as the upper left
corner and the full size of the area respectively. However, the
functions REGION-{CONTAINS,OVERLAPS}-OBJECT-P pass the object center
and the half size of the object.

Since REGION-{CONTAINS,OVERLAPS}-OBJECT-P are the only callers of
REGION-{CONTAINS,OVERLAPS}-AREA-P, adjust
REGION-{CONTAINS,OVERLAPS}-AREA-P to treat the passed location and
size as the center and half size respectively.

The problem was demonstrated by the FIXED group in the
TEST-CONTAINER-GENERIC test: The A object has a bounding region
[-5,5]x[-5,5] but the region of the root of the computed quadtree was
[0,400]x[0,400].