BSP: fixup: fix CC-MESH -> tri mesh conversion
In the case where the first 3 verts in a CC-MESH vert face all lay on a
line, we wouldn't correctly generate a face normal, which broke the rest
of the trimesh generation for that face.

This change tries all tris made from contiguous verts of the face to
produce a normal.