Solid leaf BSP (#8)
* Add BSP container

* BSP: Fix accuracy issues, and optimize BSP building

* Line/plane intersection FP accuracy was an issue, endpoint of line
  segment didn't perfectly match t=1 intersection

* Removed assertions to ensure tri verts lay on the plane - this is not
  possible to assert for small triangles, as the cross product becomes
  inaccurate. This doesn't pose a problem for BSP building, only these
  asserts.

* CC-MESH building was prohibitively slow, extra inline + type
  annotations fix this

* BSP: Ray, AABB, and serialization tests

* BSP: fixup: Remove a bad test

This test fails depending on the order of BSP construction because :funny-house is not
manifold - this failure should be allowed

* BSP: fixup: Better AABB test failure description

* BSP: Smarter picking of splitting planes

* BSP: fixup: fix CC-MESH -> tri mesh conversion

In the case where the first 3 verts in a CC-MESH vert face all lay on a
line, we wouldn't correctly generate a face normal, which broke the rest
of the trimesh generation for that face.

This change tries all tris made from contiguous verts of the face to
produce a normal.

* Add mesh coercion

* BSP: fixup: Clear the build state after building the BSP to allow mesh data to be GCed

* BSP: fixup: AABB-RESULT for aabb tests

---------

Co-authored-by: Yukari Hafner <shinmera@tymoon.eu>